Update Status: In Progress (part 1)
The file tab has been Updated, featuring a mode button that allows users to toggle between adding and removing custom applications within the Sage 300 GRP file. Currently, this functionality is available for testing only in the Accounts Payable (AP) module. Please note that object descriptions and ID values are not yet editable. Additionally, the Parent is being hardcoded to the hex arrays as they are identified and tested. please say tuned, Should have another update in the coming days.